Struct google_cloudtasks2::Queue

pub struct Queue {
    pub state: Option<String>,
    pub purge_time: Option<String>,
    pub name: Option<String>,
    pub rate_limits: Option<RateLimits>,
    pub app_engine_routing_override: Option<AppEngineRouting>,
    pub retry_config: Option<RetryConfig>,
}

A queue is a container of related tasks. Queues are configured to manage how those tasks are dispatched. Configurable properties include rate limits, retry options, queue types, and others.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations queues pause projects (response)

locations queues resume projects (response)

locations queues patch projects (request|response)

locations queues create projects (request|response)

locations queues get projects (response)

locations queues purge projects (response)

Fields
state: Option<String>

Output only. The state of the queue.

state can only be changed by called PauseQueue, ResumeQueue, or uploading queue.yaml/xml. UpdateQueue cannot be used to change state.

purge_time: Option<String>

Output only. The last time this queue was purged.

All tasks that were created before this time were purged.

A queue can be purged using PurgeQueue, the App Engine Task Queue SDK, or the Cloud Console.

Purge time will be truncated to the nearest microsecond. Purge time will be unset if the queue has never been purged.

name: Option<String>

Caller-specified and required in CreateQueue, after which it becomes output only.

The queue name.

The queue name must have the following format: projects/PROJECT_ID/locations/LOCATION_ID/queues/QUEUE_ID

PROJECT_ID can contain letters ([A-Za-z]), numbers ([0-9]), hyphens (-), colons (:), or periods (.). For more information, see Identifying projects

LOCATION_ID is the canonical ID for the queue's location. The list of available locations can be obtained by calling ListLocations. For more information, see https://cloud.google.com/about/locations/.

QUEUE_ID can contain letters ([A-Za-z]), numbers ([0-9]), or hyphens (-). The maximum length is 100 characters.

rate_limits: Option<RateLimits>

Rate limits for task dispatches.

rate_limits and retry_config are related because they both control task attempts. However they control task attempts in different ways:

rate_limits controls the total rate of dispatches from a queue (i.e. all traffic dispatched from the queue, regardless of whether the dispatch is from a first attempt or a retry).

retry_config controls what happens to particular a task after its first attempt fails. That is, retry_config controls task retries (the second attempt, third attempt, etc).

The queue's actual dispatch rate is the result of:

Number of tasks in the queue

User-specified throttling: rate_limits, retry_config, and the queue's state.

System throttling due to 429 (Too Many Requests) or 503 (Service Unavailable) responses from the worker, high error rates, or to smooth sudden large traffic spikes.

app_engine_routing_override: Option<AppEngineRouting>

Overrides for task-level app_engine_routing. These settings apply only to App Engine tasks in this queue.

If set, app_engine_routing_override is used for all App Engine tasks in the queue, no matter what the setting is for the task-level app_engine_routing.

retry_config: Option<RetryConfig>

Settings that determine the retry behavior.

For tasks created using Cloud Tasks: the queue-level retry settings apply to all tasks in the queue that were created using Cloud Tasks. Retry settings cannot be set on individual tasks.

For tasks created using the App Engine SDK: the queue-level retry settings apply to all tasks in the queue which do not have retry settings explicitly set on the task and were created by the App Engine SDK. See App Engine documentation.